orchestration/dlt/requirements.txt
dlt[filesystem,parquet]>=1.0
requests>=2.28
python-dotenv>=1.0
